+++ src/App.css
@@ -77,6 +77,11 @@
   z-index: 1000;
 }
 
+.form-header {
+  margin-top: 0;
+  color: #3333aa;
+}
+
 .formField {
   margin: 10px 0;
   padding: 10px;
@@ -88,10 +93,11 @@
 
 .formButton {
   padding: 10px 15px;
-  margin: 5px;
+  margin-right: 10px;
   border: none;
   border-radius: 5px;
   cursor: pointer;
+  margin-top: 10px;
 }
 
 .formSaveButton {
@@ -102,6 +108,11 @@
 .formCancelButton {
   background-color: #6c757d;
   color: white;
+}
+
+.formLabel {
+  font-weight: bold;
+  color: #6c757d;
 }
 
 .overlay {
