+++ src/App.js
@@ -1,6 +1,8 @@
-import React, { useState, useEffect } from 'react';
-import axios from 'axios';
 import './App.css';
+import React, { useState, useEffect } from 'react';
+import { getBooksFromApi, newBookToApi, putBookToApi, deleteBookFromApi } from './utils/Client.js';
+import BookForm from './components/Popup.js';
+import Table from './components/Table.js';
 
 const App = () => {
   const [books, setBooks] = useState([]);
@@ -19,12 +21,12 @@ const App = () => {
   }, [filters]);
 
   const fetchBooks = async () => {
-    const response = await axios.get('http://localhost:8000/api/books', { params: filters });
+    const response = await getBooksFromApi(filters);
     setBooks(response.data.data);
   };
 
   const handleDelete = async (id) => {
-    await axios.delete(`http://localhost:8000/api/books/${id}`);
+    await deleteBookFromApi(id);
     fetchBooks();
   };
 
@@ -35,9 +37,9 @@ const App = () => {
 
   const handleFormSubmit = async (bookData) => {
     if (editBook) {
-      await axios.put(`http://localhost:8000/api/books/${editBook.id}`, bookData);
+      await putBookToApi(editBook.id, bookData)
     } else {
-      await axios.post('http://localhost:8000/api/books', bookData);
+      await newBookToApi(bookData);
     }
     fetchBooks();
     setShowPopup(false);
@@ -49,7 +51,7 @@ const App = () => {
     if (updatedQuantity < 0) return;
 
     const updatedBook = { ...book, quantity: updatedQuantity };
-    await axios.put(`http://localhost:8000/api/books/${book.id}`, updatedBook);
+    await putBookToApi(book.id, updatedBook)
     fetchBooks();
   };
 
@@ -67,181 +69,27 @@ const App = () => {
       {showPopup && (
         <>
           <div className="overlay"></div>
-          <BookForm book={editBook} onSubmit={handleFormSubmit} onCancel={() => setShowPopup(false)} />
+          <BookForm
+            book={editBook}
+            onSubmit={handleFormSubmit}
+            onCancel={() => setShowPopup(false)}
+          />
         </>
       )}
-      <table className="table">
-        <thead>
-          <tr>
-            <th className="tableHeader">
-              Title<br />
-              <input
-                type="text"
-                name="title"
-                value={filters.title}
-                onChange={handleFilterChange}
-                className="filterInput"
-              />
-            </th>
-            <th className="tableHeader">
-              Description<br />
-              <input
-                type="text"
-                name="description"
-                value={filters.description}
-                onChange={handleFilterChange}
-                className="filterInput"
-              />
-            </th>
-            <th className="tableHeader">
-              Author<br />
-              <input
-                type="text"
-                name="author"
-                value={filters.author}
-                onChange={handleFilterChange}
-                className="filterInput"
-              />
-            </th>
-            <th className="tableHeader">
-              Price<br />
-              <input
-                type="number"
-                name="price"
-                min="0"
-                step="0.01"
-                value={filters.price}
-                onChange={handleFilterChange}
-                className="filterInput"
-              />
-            </th>
-            <th className="tableHeader">
-              Quantity<br />
-              <input
-                type="number"
-                name="quantity"
-                min="0"
-                step="1"
-                value={filters.quantity}
-                onChange={handleFilterChange}
-                className="filterInput"
-              />
-            </th>
-            <th className="tableHeader">Actions</th>
-          </tr>
-        </thead>
-        <tbody>
-          {books.map((book) => (
-            <tr key={book.id}>
-              <td className="tableCell">{book.title}</td>
-              <td className="tableCell">{book.description}</td>
-              <td className="tableCell">{book.author}</td>
-              <td className="tableCell">${book.price}</td>
-              <td className="tableCell">
-                <div style={{ display: 'flex', alignItems: 'center' }}>
-                  {book.quantity}
-                  <button className="button" style={{ marginLeft: '10px' }} onClick={() => handleQuantityChange(book, -1)}>-</button>
-                  <button className="button" onClick={() => handleQuantityChange(book, 1)}>+</button>
-                </div>
-              </td>
-              <td className="tableCell">
-                <button className="button deleteButton actionButton" onClick={() => handleDelete(book.id)}>Delete</button>
-                <button className="button actionButton" onClick={() => handleEdit(book)}>Edit</button>
-              </td>
-            </tr>
-          ))}
-        </tbody>
-      </table>
+      <Table
+        books={books}
+        filters={filters}
+        handleFilterChange={handleFilterChange}
+        handleQuantityChange={handleQuantityChange}
+        handleEdit={handleEdit}
+        handleDelete={handleDelete}
+      >
+      </Table>
       {!Boolean(books.length) && (
         <h2 className="noData">No data found</h2>
       )}
     </div>
   );
-
-
-};
-
-const BookForm = ({ book, onSubmit, onCancel }) => {
-  const [formData, setFormData] = useState({
-    title: book?.title || '',
-    description: book?.description || '',
-    author: book?.author || '',
-    price: book?.price || '',
-    quantity: book?.quantity || '',
-  });
-
-  const handleChange = (e) => {
-    const { name, value } = e.target;
-    setFormData((prevData) => ({
-      ...prevData,
-      [name]: value,
-    }));
-  };
-
-  const handleSubmit = (e) => {
-    e.preventDefault();
-    onSubmit(formData);
-  };
-
-  return (
-    <form className="form" onSubmit={handleSubmit}>
-      <h1 className="header">{book ? 'Edit Book' : 'Add New Book'}</h1>
-      <label htmlFor="title" className="formLabel">Title</label>
-      <input
-        id="title"
-        name="title"
-        value={formData.title}
-        onChange={handleChange}
-        placeholder="Title"
-        className="formField"
-      />
-      <label htmlFor="description" className="formLabel">Description</label>
-      <input
-        id="description"
-        name="description"
-        value={formData.description}
-        onChange={handleChange}
-        placeholder="Description"
-        className="formField"
-      />
-      <label htmlFor="author" className="formLabel">Author</label>
-      <input
-        id="author"
-        name="author"
-        value={formData.author}
-        onChange={handleChange}
-        placeholder="Author"
-        className="formField"
-      />
-      <label htmlFor="price" className="formLabel">Price</label>
-      <input
-        id="price"
-        name="price"
-        type="number"
-        step="0.01"
-        min="0"
-        value={formData.price}
-        onChange={handleChange}
-        placeholder="Price"
-        className="formField"
-      />
-      <label htmlFor="quantity" className="formLabel">Quantity</label>
-      <input
-        id="quantity"
-        name="quantity"
-        type="number"
-        step="1"
-        min="0"
-        value={formData.quantity}
-        onChange={handleChange}
-        placeholder="Quantity"
-        className="formField"
-      />
-      <button type="submit" className="formButton formSaveButton">Save</button>
-      <button type="button" onClick={onCancel} className="formButton formCancelButton">Cancel</button>
-    </form>
-  );
-
 };
 
 export default App;
